physical-ai-book/agents/requirements.txt
# Agents Requirements
openai==1.7.2
python-dotenv==1.0.0
pydantic==2.5.3
